# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ATHENA Platform API
  description: 'Production REST API for the ATHENA Decision Intelligence Platform — a single-user, advisory-only
    NSE/BSE swing-trading system. ATHENA never places or executes trades; this API is read/inspect/annotate
    only (journal entries and manually-logged fills are the sole write paths that touch the owner''s own
    record-keeping, never a broker order).


    Every endpoint requires authentication (Bearer access token from /api/v1/auth/login) except /api/v1/auth/login
    itself, /api/v1/auth/status, and the platform health/discovery endpoints. Errors follow RFC 7807 Problem
    Details. Collection endpoints share one envelope: {status, data, meta, pagination}.'
  version: 0.1.0
paths:
...
  /api/v1/decisions:
    get:
      tags:
      - Decisions
      summary: List trading decisions
      description: Retrieve a paginated collection of trading decisions, supporting sorting and filtering.
      operationId: listDecisions
      security:
      - HTTPBearer: []
      - APIKeyHeader: []
      parameters:
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: from_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: From Date
      - name: to_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: To Date
      - name: instrument_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Instrument Id
      - name: decision_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Decision Type
      - name: direction
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Direction
      - name: sort_by
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Sort By
      - name: sort_dir
        in: query
        required: false
        schema:
          enum:
          - asc
          - desc
          type: string
          default: desc
          title: Sort Dir
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Page Size
      - name: next_cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AthenaResponse_list_DecisionDTO__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/decisions/{decision_id}:
    get:
      tags:
      - Decisions
      summary: Get decision details
      description: Retrieve detailed analysis, quality gates, and trade plans for a specific decision.
      operationId: getDecision
      security:
      - HTTPBearer: []
      - APIKeyHeader: []
      parameters:
      - name: decision_id
        in: path
        required: true
        schema:
          type: string
          title: Decision Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AthenaResponse_DecisionDTO_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/decisions/{decision_id}/depth:
    get:
      tags:
      - Decisions
      summary: Get persisted decision analytical depth
      description: Render eligibility, score, confidence, and risk artifacts without recomputation.
      operationId: getDecisionDepth
      security:
      - HTTPBearer: []
      - APIKeyHeader: []
      parameters:
      - name: decision_id
        in: path
        required: true
        schema:
          type: string
          title: Decision Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AthenaResponse_DecisionDepthDTO_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/decisions/{decision_id}/counterfactual:
    get:
      tags:
      - Decisions
      summary: Get the exact quantified distance from this decision to the TRADE gate
      description: 'Arithmetic over already-persisted score/confidence/risk values and

        current config thresholds — never a recomputed decision (M-X2).'
      operationId: getDecisionCounterfactual
      security:
      - HTTPBearer: []
      - APIKeyHeader: []
      parameters:
      - name: decision_id
        in: path
        required: true
        schema:
          type: string
          title: Decision Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AthenaResponse_DecisionCounterfactualDTO_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/decisions/{decision_id}/plan-freshness:
    get:
      tags:
      - Decisions
      summary: Get the deterministic decay clock for this decision's TradePlan validity window
      description: 'Pure arithmetic over the plan''s already-persisted valid_from/valid_until

        and an as_of instant — never a recomputed plan (M-X3).'
      operationId: getDecisionPlanFreshness
      security:
      - HTTPBearer: []
      - APIKeyHeader: []
      parameters:
      - name: decision_id
        in: path
        required: true
        schema:
          type: string
          title: Decision Id
      - name: as_of
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: As Of
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AthenaResponse_TradePlanFreshnessDTO_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/decisions/{decision_id}/intraday-intelligence:
    get:
      tags:
      - Decisions
      summary: 'Get the Owner-facing Intraday Plan (ID-11): qualification, actionability, sizing, and live
        supervision'
      description: 'Composes the already-frozen EntryQualification -> EntryActionability

        -> PositionSizing -> LivePlanSupervision chain for this exact Decision,

        read-only -- never a recomputed Decision, never a full validation cycle

        (ID-11). This is a CURRENT Owner-facing read: it has no caller-selected

        historical checkpoint -- there is deliberately no `as_of`/`checkpoint`/

        `replay` query parameter -- and always evaluates against the service''s

        own current read-time instant.'
      operationId: getDecisionIntradayIntelligence
      security:
      - HTTPBearer: []
      - APIKeyHeader: []
      parameters:
      - name: decision_id
        in: path
        required: true
        schema:
          type: string
          title: Decision Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AthenaResponse_IntradayIntelligenceDTO_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/decisions/{decision_id}/context:
    get:
      tags:
      - Decisions
      summary: Get session/calendar, regime/market-health, and curated links for a decision
      description: 'Render session/calendar context, persisted regime/market-health, and

        owner-curated external links. No news ingestion, no generated rationale.'
      operationId: getDecisionContext
      security:
      - HTTPBearer: []
      - APIKeyHeader: []
      parameters:
      - name: decision_id
        in: path
        required: true
        schema:
          type: string
          title: Decision Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AthenaResponse_DecisionContextDTO_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/decisions/{decision_id}/analogs:
    get:
      tags:
      - Decisions
      summary: Get historical decisions with a similar score/confidence/risk fingerprint
      description: 'Deterministic nearest-neighbor retrieval over persisted decision

        history, with each match''s logged human response and realized outcome

        if any. No generated text, no recomputation of any comparison (M-X1).'
      operationId: getDecisionAnalogs
      security:
      - HTTPBearer: []
      - APIKeyHeader: []
      parameters:
      - name: decision_id
        in: path
        required: true
        schema:
          type: string
          title: Decision Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 20
          minimum: 1
          default: 5
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AthenaResponse_DecisionAnalogsDTO_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/decisions/{decision_id}/journal:
    post:
      tags:
      - Decisions
      summary: Record the owner's response to a decision (accept/reject/ignore)
      description: Persist the owner's response (M-X0, R-9) — nothing is unrecorded.
      operationId: recordDecisionJournalEntry
      security:
      - HTTPBearer: []
      - APIKeyHeader: []
      parameters:
      - name: decision_id
        in: path
        required: true
        schema:
          type: string
          title: Decision Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordJournalRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AthenaResponse_JournalEntryDTO_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Decisions
      summary: Get the owner's most recent response to a decision
      description: None if the owner has never recorded a response for this decision.
      operationId: getDecisionJournalEntry
      security:
      - HTTPBearer: []
      - APIKeyHeader: []
      parameters:
      - name: decision_id
        in: path
        required: true
        schema:
          type: string
          title: Decision Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AthenaResponse_Union_JournalEntryDTO__NoneType__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/decisions/{decision_id}/trace:
    get:
      tags:
      - Decisions
      summary: Get decision execution trace DAG flow
      description: Retrieve the reasoning DAG trace showing pipeline stages from ingest to safety checks.
      operationId: getDecisionTrace
      security:
      - HTTPBearer: []
      - APIKeyHeader: []
      parameters:
      - name: decision_id
        in: path
        required: true
        schema:
          type: string
          title: Decision Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AthenaResponse_DecisionTraceDTO_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AthenaResponse_DecisionDTO_:
      properties:
        status:
          type: string
          const: success
          title: Status
          default: success
        data:
          $ref: '#/components/schemas/DecisionDTO'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
        pagination:
          anyOf:
          - $ref: '#/components/schemas/PaginationMeta'
          - type: 'null'
        links:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Links
      type: object
      required:
      - data
      - meta
      title: AthenaResponse[DecisionDTO]
...
    AthenaResponse_list_DecisionDTO__:
      properties:
        status:
          type: string
          const: success
          title: Status
          default: success
        data:
          items:
            $ref: '#/components/schemas/DecisionDTO'
          type: array
          title: Data
        meta:
          $ref: '#/components/schemas/ResponseMeta'
        pagination:
          anyOf:
          - $ref: '#/components/schemas/PaginationMeta'
          - type: 'null'
        links:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Links
      type: object
      required:
      - data
      - meta
      title: AthenaResponse[list[DecisionDTO]]
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
servers:
- url: http://127.0.0.1:8000
  description: Local workstation (ATHENA is localhost-only by design; never deployed remotely)